+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Help text for the `--assign-to-me` flag
 
 ## [0.3.0] - 2024-04-17
 ### Added
@@ -49,4 +51,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.1.0]: https://github.com/nekeal/issx/compare/0.0.4...0.1.0
 [0.0.4]: https://github.com/nekeal/issx/compare/0.0.2...0.0.4
 [0.0.2]: https://github.com/nekeal/issx/tree/0.0.2
-