+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.3] - 2024-04-29
 ### Fixed
 - Fix printing toml config in the wizard
 
@@ -78,7 +80,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Service for copying issues between projects
 - CLI for copying issues and verifying auth configuration with Gitlab ([b7eed84](https://github.com/nekeal/issx/commit/b7eed844239f0c251c9501a0c455ab457c4ed910))
 
-[Unreleased]: https://github.com/nekeal/issx/compare/0.5.2...master
+[Unreleased]: https://github.com/nekeal/issx/compare/0.5.3...master
+[0.5.3]: https://github.com/nekeal/issx/compare/0.5.2...0.5.3
 [0.5.2]: https://github.com/nekeal/issx/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/nekeal/issx/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/nekeal/issx/compare/0.4.0...0.5.0
@@ -90,3 +93,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.1.0]: https://github.com/nekeal/issx/compare/0.0.4...0.1.0
 [0.0.4]: https://github.com/nekeal/issx/compare/0.0.2...0.0.4
 [0.0.2]: https://github.com/nekeal/issx/tree/0.0.2
+